+++ examples/complete/main.tf
@@ -1,19 +1,14 @@
-#### NOTE: This example should be run on management account
-
 locals {
-  name = "complete-org-trail-example"
+  name = "complete-boldlink-example"
 }
 
-module "org_cloudtrail" {
+module "complete" {
   source                     = "../../"
   name                       = local.name
   enable_log_file_validation = true
   bucket_name                = local.name
   trail_name                 = local.name
-  is_organization_trail      = true
-  protect_cloudtrail         = true ## Protects against disabling cloudtrail
   enable_logging             = true
-
   event_selectors = [
     {
       read_write_type = "All"