+++ examples/complete/outputs.tf
@@ -1,6 +1,1 @@
-output "outputs" {
-  value = [
-    module.org_cloudtrail,
-  ]
-  description = "Various output values for the created resources"
-}
+#(empty)