+++ examples/non_org_trail_with_external_bucket/outputs.tf
@@ -1,6 +1,0 @@
-output "outputs" {
-  value = [
-    module.aws_cloudtrail,
-  ]
-  description = "The various values for the created resources"
-}