+++ examples/organization_trail_with_external_bucket/main.tf
@@ -1,5 +1,7 @@
+### **NOTE**: This example should be run on management account
+
 locals {
-  name            = "org-trail-bucket-example"
+  name            = "organizational-boldlink-example"
   account_id      = data.aws_caller_identity.current.account_id
   region          = data.aws_region.current.name
   organization_id = data.aws_organizations_organization.current.id